+++ includes/SpeechoidConnector.php
@@ -18,11 +18,21 @@ class SpeechoidConnector {
 	/** @var string Speechoid URL, without trailing slash. */
 	private $url;
 
-	public function __construct() {
-		$config = MediaWikiServices::getInstance()
-			->getConfigFactory()
-			->makeConfig( 'wikispeech' );
+	/** @var int Default timeout awaiting HTTP response in seconds. */
+	private $defaultHttpResponseTimeoutSeconds;
+
+	/**
+	 * @param Config $config
+	 * @since 0.1.5
+	 */
+	public function __construct( $config ) {
 		$this->url = rtrim( $config->get( 'WikispeechSpeechoidUrl' ), '/' );
+
+		if ( $config->get( 'WikispeechSpeechoidResponseTimeoutSeconds' ) ) {
+			$this->defaultHttpResponseTimeoutSeconds = intval(
+				$config->get( 'WikispeechSpeechoidResponseTimeoutSeconds' )
+			);
+		}
 	}
 
 	/**
@@ -32,24 +42,31 @@ class SpeechoidConnector {
 	 * @param string $language
 	 * @param string $voice
 	 * @param string $text
+	 * @param int|null $responseTimeoutSeconds Seconds before timing out awaiting response.
+	 *  Falsy value defaults to config value WikispeechSpeechoidResponseTimeoutSeconds,
+	 *  which if falsy (e.g. 0) defaults to MediaWiki default.
 	 * @return array Response from Speechoid, parsed as associative array.
 	 * @throws SpeechoidConnectorException On Speechoid I/O errors.
 	 */
 	public function synthesize(
 		$language,
 		$voice,
-		$text
+		$text,
+		$responseTimeoutSeconds = null
 	): array {
 		$requestFactory = MediaWikiServices::getInstance()->getHttpRequestFactory();
-		$requestParameters = [
+		$postData = [
 			'lang' => $language,
 			'voice' => $voice,
 			'input' => $text
 		];
-		$responseString = $requestFactory->post(
-			$this->url,
-			[ 'postData' => $requestParameters ]
-		);
+		$options = [ 'postData' => $postData ];
+		if ( $responseTimeoutSeconds ) {
+			$options['timeout'] = $responseTimeoutSeconds;
+		} elseif ( $this->defaultHttpResponseTimeoutSeconds ) {
+			$options['timeout'] = $this->defaultHttpResponseTimeoutSeconds;
+		}
+		$responseString = $requestFactory->post( $this->url, $options );
 		if ( !$responseString ) {
 			throw new SpeechoidConnectorException( 'Unable to communicate with Speechoid.' );
 		}